+++ src/onegov.activity/HISTORY.rst
@@ -1,6 +1,9 @@
 Changelog
 ---------
 
+- Adds the ability to get the set of used activity tags.
+  [href]
+
 0.0.1 (2016-09-22)
 ~~~~~~~~~~~~~~~~~~~